+++ packages/app/src/languages/actions/language-actions.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { ProgrammingLanguage } from "shared-types";
-import { ApiUrl } from "../../common/utils/api-url-config";
+import { ApiUrl } from "../../common/utils/backend-url";
 import { LanguagesStore } from "../stores/languages-store";
 
 export const loadAllLanguages = async () => {